# doc/api.yaml
openapi: 3.0.0
servers:
  - url: http://localhost:3000/api/v3
info:
  title: WASAText
  description: Web And Software Architecture Exam Project
  version: 1.0.0
security:
  - bearerAuth: []
...
paths:
  /groups:
    post:
      tags: ["groups"]
      summary: Create a group
      description: Create a group
      operationId: createGroup
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              description: The group data
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/Name"
                photo:
                  $ref: "#/components/schemas/Photo"
      responses:
        '201':
          description: Created
          content:
            application/json:
              example:
                type: object
              schema:
                $ref: "#/components/schemas/Group"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '409':
          $ref: "#/components/responses/Conflict"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /groups/{groupId}:
    parameters:
      - name: groupId
        in: path
        description: Group id
        required: true
        schema:
          type: integer
          format: int64
    post:
      tags: ["groups"]
      summary: Add to group
      description: Add to group
      operationId: addToGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: The user to add
              type: object
              properties:
                userId:
                  description: User id
                  type: integer
                  format: int64
      responses:
        '201':
          description: Created
          content:
            application/json:
              example:
                type: object
              schema:
                description: The user added to the group
                type: object
                properties:
                  userId:
                    description: User id
                    type: integer
                    format: int64
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
    get:
      tags: ["groups"]
      summary: Get group data
      description: Get group data
      operationId: getGroup
      responses:
        '200':
          description: Successful
          content:
            application/json:
              example:
                type: object
              schema:
                $ref: "#/components/schemas/Group"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: ["groups"]
      summary: Delete group
      description: Delete group
      operationId: deleteGroup
      responses:
        '204':
          description: No Content
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /groups/{groupId}/name:
    parameters:
      - name: groupId
        in: path
        description: Group id
        required: true
        schema:
          type: integer
          format: int64
    put:
      tags: ["groups"]
      operationId: setGroupName
      summary: Change group name
      description: Change group name
      requestBody:
        required: true
        content:
          application/json:
            example:
              type: object
            schema:
              description: The group name
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/Name"
      responses:
        '200':
          description: Successful
          content:
            application/json:
              example:
                type: object
              schema:
                description: The group name
                type: object
                properties:
                  name:
                    $ref: "#/components/schemas/Name"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /groups/{groupId}/photo:
    parameters:
      - name: groupId
        in: path
        description: Group id
        required: true
        schema:
          type: integer
          format: int64
    put:
      tags: ["groups"]
      operationId: setGroupPhoto
      summary: Update group photo
      description: Update group photo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              description: The group photo
              type: object
              properties:
                photo:
                  $ref: "#/components/schemas/Photo"
      responses:
        '200':
          description: Successful
          content:
            application/json:
              example:
                type: object
              schema:
                description: The photo url
                type: object
                properties:
                  photoUrl:
                    $ref: "#/components/schemas/PhotoUrl"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /groups/{groupId}/message:
    parameters:
      - name: groupId
        in: path
        description: Group id
        required: true
        schema:
          type: integer
          format: int64
    post:
      tags: ["groups"]
      summary: Send a message to a group
      description: Send a message to a group
      operationId: sendMessageToGroup
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MessageContent"
      responses:
        '201':
          description: Created
          content:
            application/json:
              example:
                type: object
              schema:
                $ref: "#/components/schemas/Message"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /groups/{groupId}/fmessage:
    parameters:
      - name: groupId
        in: path
        description: Group id
        required: true
        schema:
          type: integer
          format: int64
    post:
      tags: ["groups"]
      summary: Forward a message to a group
      description: Forward a message to a group
      operationId: forwardMessageToGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: The message to forward
              type: object
              properties:
                messageId:
                  description: Message id
                  type: integer
                  format: int64
      responses:
        '201':
          description: Created
          content:
            application/json:
              example:
                type: object
              schema:
                $ref: "#/components/schemas/Message"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /groups/{groupId}/user:
    parameters:
      - name: groupId
        in: path
        description: Group id
        required: true
        schema:
          type: integer
          format: int64
    delete:
      tags: ["groups"]
      summary: Leave the group
      description: Leave the group
      operationId: leaveGroup
      responses:
        '204':
          description: No Content
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /groups/{groupId}/user/{userId}:
    parameters:
      - name: groupId
        in: path
        description: Group id
        required: true
        schema:
          type: integer
          format: int64
      - name: userId
        in: path
        description: User id
        required: true
        schema:
          type: integer
          format: int64
    delete:
      tags: ["groups"]
      summary: Remove user from group
      description: Remove user from group
      operationId: removeUser
      responses:
        '204':
          description: No Content
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
  /messages/{messageId}:
    parameters:
      - name: messageId
        in: path
        description: Message id
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: ["messages"]
      operationId: getMessage
      summary: Get message data
      description: Get message data
      responses:
        '200':
          description: Successful
          content:
            application/json:
              example:
                type: object
              schema:
                $ref: "#/components/schemas/Message"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: ["messages"]
      operationId: deleteMessage
      summary: Delete message
      description: Delete message
      responses:
        '204':
          description: No Content
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
  schemas:
...
    Group:
      description: Group
      type: object
      properties:
        conversationId:
          description: Group conversation id
          type: integer
          format: int64
        founderId:
          description: Group founder id
          type: integer
          format: int64
        name:
          $ref: "#/components/schemas/Name"
        photoUrl:
          $ref: "#/components/schemas/PhotoUrl"
        createdAt:
          description: Group timestamp
          type: string
          format: date-time
    Name:
      description: Generic name
      type: string
      pattern: '^.*$'
      minLength: 1
      maxLength: 1024
    PhotoUrl:
      description: Photo URL
      type: string
      format: url
      minLength: 1
      maxLength: 1024
    Message:
      description: Message
      type: object
      properties:
        messageId:
          description: Message id
          type: integer
          format: int64
        text:
          description: Message text
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 1024
        senderId:
          description: Sender id
          type: integer
          format: int64
        conversationId:
          description: Conversation id
          type: integer
          format: int64
        createdAt:
          description: Message timestamp
          type: string
          format: date-time
        isForwarded:
          description: Message is forwarded
          type: boolean
        commentTo:
          description: Commented message id
          type: integer
          format: int64
        attachmentUrl:
          description: Attachment URL
          type: string
          format: url
        mediaType:
          $ref: "#/components/schemas/MediaType"
    MessageContent:
      description: Message content
      type: object
      properties:
        text:
          description: Message text
          type: string
          minLength: 0
          maxLength: 1024
          pattern: "^.*$"
        file:
          description: Binary file
          type: string
          format: binary
          minLength: 1
          maxLength: 1024
        mediaType:
          $ref: "#/components/schemas/MediaType"
    MediaType:
      description: Media type
      type: string
      enum: [image, video, audio, file]
...
    Photo:
      description: Photo
      type: string
      format: binary
      minLength: 1
      maxLength: 1014
...